First running version of the IDL documentation scripts for Doxygen
Ruby script has been modified so that it parses a single file (i.e does
not look for all the files recursively, this is Doxygen task). The
Doxygen configuration file has been updated so that it can be used as a
template. The sample programs (i.e. pro files) have been extended with
more functions/procedures.